+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v0.3.7 (11/09/2026)
 
 - Remove `<TargetFramework>net10.0</TargetFramework>` from WixInstaller template
+- Add `UnixSystem.SetUnixPermissions(...)` to set the permissions of a file or directory
+- Add `UnixSystem.ChangeUnixPermissions(...)` to change the current permissions of a file or directory
 - Add `StageKitBuild.UnixFilePermissions` to apply configured `chmod` modes to validated files beneath each Unix-like
   runtime's publish directory after publishing
 - Improve the `StageKitBuild.SoftwareName` property to guess from more properties